chore: §6-2 중복 타입 통합
