+++ apps/web/src/app/archive/wish/[id]/_types/wish.ts
@@ -23,11 +23,4 @@ export type GetWishResponseT = {
   );
 };
 
-export type PatchWishRequestT = {
-  name: string;
-  currentPrice: number;
-  image: File;
-  currency: string;
-};
-
 export type PatchWishResponseT = GetWishResponseT;
