+++ apps/web/src/app/notification/_apis/postNotificationsRead.ts
@@ -2,7 +2,7 @@ import { clientApi } from '@/apis/client';
 import { ENDPOINTS } from '@/consts/api';
 import type { ApiResponseT } from '@/types/api';
 
-import type { NotificationCategoryT } from '../types/notification';
+import type { NotificationCategoryT } from '@/types/notification';
 
 type PostNotificationsReadResponseT = ApiResponseT<{
   unreadCount: number;
