+++ apps/web/src/app/tournament/[id]/item/[itemId]/_types/tournamentItem.ts
@@ -24,11 +24,4 @@ export type GetTournamentItemResponseT = {
     }
 );
 
-export type PatchTournamentItemRequestT = {
-  name: string;
-  currentPrice: number;
-  image: File;
-  currency: string;
-};
-
 export type PatchTournamentItemResponseT = GetTournamentItemResponseT;
