+++ apps/web/src/types/notification.ts
@@ -12,6 +12,29 @@ export type NotificationTypeT =
 
 export type NotificationKindT = 'WISH' | 'TOURNAMENT';
 
+export type NotificationCategoryT = 'ACTIVITY' | 'SYSTEM';
+
+export type NotificationItemT = {
+  id: number;
+  type: NotificationTypeT;
+  category: NotificationCategoryT;
+  title: string;
+  body: string;
+  imageUrl: string;
+  refId: number;
+  isRead: boolean;
+  createdAt: string;
+  kind?: NotificationKindT;
+  tournamentId?: number;
+  tournamentItemId?: number;
+};
+
+export type NotificationListDataT = {
+  items: NotificationItemT[];
+  unreadCount: number;
+  unreadCountByCategory: Record<NotificationCategoryT, number>;
+};
+
 export type SilentSyncSsePayloadT =
   | {
       type: 'TOURNAMENT_ITEM_PARSED';
